# AVX2 alpha blitters