--- a/src_c/alphablit.c
+++ b/src_c/alphablit.c
@@ -220,7 +220,7 @@ SoftBlitPyGame(SDL_Surface *src, SDL_Rect *srcrect, SDL_Surface *dst,
                             src->format->Gmask == dst->format->Gmask &&
                             src->format->Bmask == dst->format->Bmask) {
 /* If our source and destination are the same ARGB 32bit
-   format we can use SSE2/NEON to speed up the blend */
+   format we can use SSE2/NEON/AVX2 to speed up the blend */
 #if PG_ENABLE_SSE_NEON
                             if ((pg_HasSSE_NEON()) && (src != dst)) {
                                 if (info.src_blanket_alpha != 255) {
@@ -232,6 +232,10 @@ SoftBlitPyGame(SDL_Surface *src, SDL_Rect *srcrect, SDL_Surface *dst,
                                             dst->format->format) &&
                                         info.dst_blend != SDL_BLENDMODE_NONE) {
                                         alphablit_alpha_sse2_argb_no_surf_alpha(
+                                            &info);
+                                    }
+                                    else if (SDL_HasAVX2()) {
+                                        alphablit_alpha_avx2_argb_no_surf_alpha_opaque_dst(
                                             &info);
                                     }
                                     else {