--- a/src_c/simd_blitters.h
+++ b/src_c/simd_blitters.h
@@ -65,6 +65,10 @@ pg_has_avx2();
 void
 alphablit_alpha_avx2_argb_no_surf_alpha_opaque_dst(SDL_BlitInfo *info);
 void
+alphablit_alpha_avx2_argb_no_surf_alpha(SDL_BlitInfo *info);
+void
+alphablit_alpha_avx2_argb_surf_alpha(SDL_BlitInfo *info);
+void
 blit_blend_rgba_mul_avx2(SDL_BlitInfo *info);
 void
 blit_blend_rgb_mul_avx2(SDL_BlitInfo *info);