Adds an AVX2 version of all 3 SSE2 alpha blitter variants.

Shares heritage with predecessor https://github.com/pygame/pygame/pull/3425, but is much better.

Approach: use masked load/store to enable stride switching using all full width AVX2 instructions to reduce code repetition. Further use macros to reduce code repetition and increase readability by leaving only the important stuff in the top level. Use alpha mask and alpha shift to allow dynamic position of the alpha channel rather than restricting the blitter to ARGB-- pulling out or inserting the alpha channel requires special care, but all operations occur in "native pixelformat space" (ofc stretched out to 16 bits per channel for calculations). Of course the source and destination need to have the same RGB channel positions still.

Benchmarking results:
```
# LARGE_SURF
# OPAQUE_DST and not GLOBAL_ALPHA_SRC: 1.61 seconds -> 1.10 seconds (45% faster)
# OPAQUE_DST and GLOBAL_ALPHA_SRC: 5.62 seconds -> 1.26 seconds (345% faster)
# not OPAQUE_DST and not GLOBAL_ALPHA_SRC: 2.79 seconds -> 1.15 seconds (140% faster)
# not OPAQUE_DST and GLOBAL_ALPHA_SRC: 5.58 seconds -> 1.30 seconds (330% faster)
```

Performance test script:
```py
import pygame
import time

# MODES
LARGE_SURF=True

OPAQUE_DST=True
GLOBAL_ALPHA_SRC=False

# OPAQUE_DST and not GLOBAL_ALPHA_SRC -> alphablit_alpha_avx2_argb_no_surf_alpha_opaque_dst
# OPAQUE_DST and GLOBAL_ALPHA_SRC -> alphablit_alpha_avx2_argb_surf_alpha
# not OPAQUE_DST and not GLOBAL_ALPHA_SRC -> alphablit_alpha_avx2_argb_no_surf_alpha
# not OPAQUE_DST and GLOBAL_ALPHA_SRC -> alphablit_alpha_avx2_argb_surf_alpha

pygame.init()

if LARGE_SURF:
    surf_size = (1920,1080)
    test_iterations = 1000
else:
    surf_size = (50,50)
    test_iterations = 100000

#screen = pygame.Surface(surf_size, 0 if OPAQUE_DST else pygame.SRCALPHA, depth=32)
screen = pygame.Surface(surf_size, depth=32)
screen.fill((60, 70, 80))

surf = pygame.Surface(surf_size, pygame.SRCALPHA, depth=32)
surf.fill((5,216,77,120))
if GLOBAL_ALPHA_SRC:
    surf.set_alpha(124)

print("== Surface information ==")
print(screen, surf)
print(screen.get_shifts(), surf.get_shifts())

start = time.time()
for _ in range(test_iterations):
    screen.blit(surf, (0, 0))    

print("== Time ==")
print(time.time() - start)
```
Edit: edited performance script and re ran benchmarks because I had OPAQUE=True mapping to using SRCALPHA, which is incorrect.