@Starbuck5 2023-03-16
I was wondering why I only got a 45% percent performance gain on the opaque dst blitter, when before I was getting a 200% improvement, and I've figured out that the new blitters are less efficient traversing vertically than the old ones. So if you reduce the height of the blit, these new AVX ones look better and better relative to the SSE ones.

I bet this can be tuned to shine more.