@MyreMylar 2023-03-19
I ran a quick experiment on using [Duff's device](https://en.wikipedia.org/wiki/Duff%27s_device) over the standard for loop you've used here, just since that is in all of our other blitters. As far as I can tell Duff's device doesn't appear to speed this up (it came out slightly slower for most blit sizes), but I only bashed it in quickly. Perhaps we should run an experiment on pulling it out of the other blitters if it is a hack that has had it's day. I guess switching this over might be why you've seen different performance characteristics depending on the height versus the width?

Anyway, I've run it though a bunch of my performance testing programs and it seems a solid all round speed boost for any alpha blits at, or above 8x8 pixels. It also smooths out a lot of the odd versus even pixel width differences we had before.